CLOUDP-135435: Fixes (#1469)
